backend/requirements.txt
# FastAPI Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
fastapi-mail==1.4.1
twilio==8.10.3
redis==5.0.1
apscheduler==3.10.4
opencv-python==4.8.1.78
ultralytics==8.0.196
numpy==1.24.3
Pillow==10.1.0
python-socketio==5.10.0
websockets==12.0
httpx==0.25.2
python-dotenv==1.0.0
email-validator==2.1.0
